src/automations/daily-summary/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
pytz==2024.1
